requirements.txt
streamlit
numpy
pillow
opencv-python-headless==4.8.0.76
supervision
roboflow==0.2.30
